+++ src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { AgroupationComponent } from './views/group-items/agroupation.component';
+import { AgroupationComponent } from './views/agroupation/agroupation.component';
 import { MainMenuComponent } from './views/main-menu/main-menu.component';
 
 const routes: Routes = [
